Creating conda environment, refactoring code base, improving system prompt, adding prompt cleaning
