--- requirements.txt
@@ -1,6 +1,0 @@
-pyyaml
-PyMuPDF
-faiss-cpu
-tqdm
-nltk
-sentence-transformers
