+++ src/__init__.py
@@ -1,0 +1,4 @@
+"""TokenSmith: Long-context document Question-Answering RAG application"""
+
+__version__ = "0.1.0"
+__author__ = "Database Group GT"
