+++ src/llama_path.txt
@@ -1,0 +1,1 @@
+/opt/homebrew/bin/llama-cli
